Abort service when token expires.
Aborting service when refresh token expires can be beneficial.  It is
an error that may require an intervention of an administrator.  It
prevents clients from busy looping due to the error.  It would give a
watchdog service a chance to restart the service, which may pop up the
browser window for re-authorisation.