Support multipart/mixed.
Outlook sends e-mail in the MIME type, "multipart/mixed" if the e-mail
contains attachments.  "multipart/mixed" seems to consist of a part for
the text of the e-mail in "multipart/alternative" followed by parts
containing the attachments.  This change includes some support for
multipart/mixed.  It reads the part containing "multipart/alternative"
and uses it to create the service response.  It drops any attachments.